requirements.txt
mysql_connector
python-dotenv
